refactor: extract camera-ready artifact writers
